fix: network input validation
